backend/requirements.txt
blinker==1.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dominate==2.5.2
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Datepicker==0.14
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.21
pandas==1.1.2
protobuf==3.13.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
sentry-sdk==0.17.6
six==1.15.0
visitor==0.1.3
humanize~=2.6.0
pyspellchecker==0.5.5
PyMySQL~=0.10.1
SQLAlchemy~=1.3.19
google-cloud-logging==1.15.1
tweepy~=3.9.0
requests-html==0.10.0
google-cloud-language==1.3.0
google-cloud-storage==1.31.0
sendgrid==6.4.7
gunicorn==20.0.4
python-dotenv==0.16.0
sshtunnel==0.4.0